refactor: add index.ts for import components
