+++ src/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import MainPage from './pages/MainPage';
 import styles from './App.module.css';
 import Lobby from './pages/Lobby/Lobby';
-import Header from './Header/Header';
+import Header from './Components/Header/Header';
 import Footer from './Components/Footer/Footer';
 
 const App: React.FC = () => {
